Update project structure, views, and shared modules
Move asset catalogs to Shared folder, refactor views into extensions,
add batch summarizer, improve article extraction, and update database
and parser logic.
