--- SakuraRSS/Assets.xcassets/ActionExtensionIcon.appiconset/Contents.json
@@ -1,14 +1,0 @@
-{
-  "images" : [
-    {
-      "filename" : "Add Feed.001.png",
-      "idiom" : "universal",
-      "platform" : "ios",
-      "size" : "1024x1024"
-    }
-  ],
-  "info" : {
-    "author" : "xcode",
-    "version" : 1
-  }
-}
