--- SakuraRSS/Assets.xcassets/SakuraIcon.imageset/Contents.json
@@ -1,15 +1,0 @@
-{
-  "images" : [
-    {
-      "filename" : "Sakura.006.png",
-      "idiom" : "universal"
-    }
-  ],
-  "info" : {
-    "author" : "xcode",
-    "version" : 1
-  },
-  "properties" : {
-    "template-rendering-intent" : "template"
-  }
-}
