--- SakuraRSS/Assets.xcassets/WidgetBackground.colorset/Contents.json
@@ -1,36 +1,0 @@
-{
-  "colors" : [
-    {
-      "color" : {
-        "color-space" : "extended-gray",
-        "components" : {
-          "alpha" : "1.000",
-          "white" : "1.000"
-        }
-      },
-      "idiom" : "universal"
-    },
-    {
-      "appearances" : [
-        {
-          "appearance" : "luminosity",
-          "value" : "dark"
-        }
-      ],
-      "color" : {
-        "color-space" : "display-p3",
-        "components" : {
-          "alpha" : "1.000",
-          "blue" : "0.118",
-          "green" : "0.110",
-          "red" : "0.110"
-        }
-      },
-      "idiom" : "universal"
-    }
-  ],
-  "info" : {
-    "author" : "xcode",
-    "version" : 1
-  }
-}
