--- SakuraRSS/Assets.xcassets/sns.twitter.x.symbolset/Contents.json
@@ -1,12 +1,0 @@
-{
-  "info" : {
-    "author" : "xcode",
-    "version" : 1
-  },
-  "symbols" : [
-    {
-      "filename" : "sns.twitter.x.svg",
-      "idiom" : "universal"
-    }
-  ]
-}
