+++ SakuraRSS/Views/More/LabsView.swift
@@ -5,6 +5,8 @@ struct LabsView: View {
     @AppStorage("Labs.XProfileFeeds") private var xProfileFeedsEnabled: Bool = false
     @AppStorage("Labs.YouTubePlayer") private var youTubePlayerEnabled: Bool = false
 
+    @State private var isXSignedIn = false
+    @State private var showXLogin = false
     @State private var isYouTubeSignedIn = false
     @State private var showYouTubeLogin = false
 
@@ -13,17 +15,31 @@ struct LabsView: View {
     }
 
     var body: some View {
-        Form {
+        List {
             Section {
                 Text("Labs.Warning \(appName)")
                     .font(.callout)
-                    .foregroundStyle(.secondary)
+                    .foregroundStyle(.primary)
+                    .listRowBackground(Color.clear)
             }
 
             Section {
                 Toggle(String(localized: "Labs.XProfileFeeds"), isOn: $xProfileFeedsEnabled)
-            } header: {
-                Text("Labs.Section.Features")
+
+                if xProfileFeedsEnabled {
+                    if isXSignedIn {
+                        Button(String(localized: "Labs.XProfileFeeds.SignOut")) {
+                            Task {
+                                await XProfileScraper.clearXSession()
+                                isXSignedIn = false
+                            }
+                        }
+                    } else {
+                        Button(String(localized: "Labs.XProfileFeeds.SignIn")) {
+                            showXLogin = true
+                        }
+                    }
+                }
             } footer: {
                 Text("Labs.XProfileFeeds.Footer")
             }
@@ -53,6 +69,13 @@ struct LabsView: View {
         .toolbarTitleDisplayMode(.inlineLarge)
         .scrollContentBackground(.hidden)
         .sakuraBackground()
+        .sheet(isPresented: $showXLogin) {
+            Task {
+                isXSignedIn = await XProfileScraper.hasXSession()
+            }
+        } content: {
+            XLoginView()
+        }
         .sheet(isPresented: $showYouTubeLogin) {
             Task {
                 isYouTubeSignedIn = await YouTubePlayerView.hasYouTubeSession()
@@ -61,6 +84,7 @@ struct LabsView: View {
             YouTubeLoginView()
         }
         .task {
+            isXSignedIn = await XProfileScraper.hasXSession()
             isYouTubeSignedIn = await YouTubePlayerView.hasYouTubeSession()
         }
     }
