+++ SakuraRSS/Views/Shared/XLoginView.swift
@@ -53,7 +53,9 @@ private struct XLoginWebView: UIViewRepresentable {
         config.websiteDataStore = .default()
         let webView = WKWebView(frame: .zero, configuration: config)
         webView.navigationDelegate = context.coordinator
+        // swiftlint:disable line_length
         webView.customUserAgent = "Mozilla/5.0 (iPhone; CPU iPhone OS 18_0 like Mac OS X) AppleWebKit/605.1.15 (KHTML, like Gecko) Version/18.0 Mobile/15E148 Safari/604.1"
+        // swiftlint:enable line_length
 
         if let loginURL = URL(string: "https://x.com/i/flow/login") {
             webView.load(URLRequest(url: loginURL))
