+++ SakuraRSS/Views/Shared/YouTubeLoginView.swift
@@ -51,11 +51,12 @@ private struct YouTubeLoginWebView: UIViewRepresentable {
         config.websiteDataStore = .default()
         let webView = WKWebView(frame: .zero, configuration: config)
         webView.navigationDelegate = context.coordinator
+        // swiftlint:disable line_length
         webView.customUserAgent = "Mozilla/5.0 (iPhone; CPU iPhone OS 18_0 like Mac OS X) AppleWebKit/605.1.15 (KHTML, like Gecko) Version/18.0 Mobile/15E148 Safari/604.1"
-
         if let loginURL = URL(string: "https://accounts.google.com/ServiceLogin?service=youtube&continue=https://m.youtube.com/") {
             webView.load(URLRequest(url: loginURL))
         }
+        // swiftlint:enable line_length
         return webView
     }
