+++ SakuraRSS/Views/Shared/YouTubePlayerView.swift
@@ -126,6 +126,14 @@ struct YouTubePlayerView: View {
         .sakuraBackground()
         .navigationBarTitleDisplayMode(.inline)
         .navigationTitle(article.title)
+        .onReceive(
+            NotificationCenter.default.publisher(for: UIDevice.orientationDidChangeNotification)
+        ) { _ in
+            let orientation = UIDevice.current.orientation
+            if orientation.isLandscape {
+                enterFullscreen()
+            }
+        }
         .task {
             let signedIn = await YouTubePlayerView.hasYouTubeSession()
             let premium = signedIn ? await YouTubePlayerView.hasYouTubePremium() : false
@@ -225,7 +233,9 @@ struct YouTubePlayerView: View {
         let config = WKWebViewConfiguration()
         config.websiteDataStore = .default()
         let webView = WKWebView(frame: .zero, configuration: config)
+        // swiftlint:disable line_length
         webView.customUserAgent = "Mozilla/5.0 (iPhone; CPU iPhone OS 18_0 like Mac OS X) AppleWebKit/605.1.15 (KHTML, like Gecko) Version/18.0 Mobile/15E148 Safari/604.1"
+        // swiftlint:enable line_length
 
         return await withCheckedContinuation { continuation in
             let delegate = PremiumCheckDelegate { isPremium in
@@ -245,260 +255,5 @@ struct YouTubePlayerView: View {
             || cookie.domain.lowercased().contains("accounts.google.com") {
             await store.httpCookieStore.deleteCookie(cookie)
         }
-    }
-}
-
-private struct YouTubePlayerWebView: UIViewRepresentable {
-
-    let urlString: String
-    @Binding var isPlaying: Bool
-    @Binding var currentTime: TimeInterval
-    @Binding var duration: TimeInterval
-    @Binding var webView: WKWebView?
-
-    func makeCoordinator() -> Coordinator {
-        Coordinator(isPlaying: $isPlaying, currentTime: $currentTime, duration: $duration)
-    }
-
-    func makeUIView(context: Context) -> WKWebView {
-        let config = WKWebViewConfiguration()
-        config.websiteDataStore = .default()
-        config.allowsInlineMediaPlayback = true
-        config.mediaTypesRequiringUserActionForPlayback = []
-        config.allowsPictureInPictureMediaPlayback = true
-
-        let webView = WKWebView(frame: .zero, configuration: config)
-        webView.navigationDelegate = context.coordinator
-        webView.scrollView.isScrollEnabled = false
-        webView.isOpaque = false
-        webView.backgroundColor = .black
-        webView.scrollView.backgroundColor = .black
-        webView.customUserAgent = "Mozilla/5.0 (iPhone; CPU iPhone OS 18_0 like Mac OS X) AppleWebKit/605.1.15 (KHTML, like Gecko) Version/18.0 Mobile/15E148 Safari/604.1"
-        webView.isUserInteractionEnabled = false
-        if let url = URL(string: urlString) {
-            webView.load(URLRequest(url: url))
-        }
-
-        DispatchQueue.main.async {
-            self.webView = webView
-        }
-
-        return webView
-    }
-
-    func updateUIView(_ uiView: WKWebView, context: Context) {}
-
-    static func dismantleUIView(_ uiView: WKWebView, coordinator: Coordinator) {
-        coordinator.invalidateObserver()
-    }
-
-    @MainActor
-    final class Coordinator: NSObject, WKNavigationDelegate {
-        @Binding var isPlaying: Bool
-        @Binding var currentTime: TimeInterval
-        @Binding var duration: TimeInterval
-        private var playbackObserver: Timer?
-
-        init(isPlaying: Binding<Bool>, currentTime: Binding<TimeInterval>, duration: Binding<TimeInterval>) {
-            _isPlaying = isPlaying
-            _currentTime = currentTime
-            _duration = duration
-        }
-
-        func invalidateObserver() {
-            playbackObserver?.invalidate()
-            playbackObserver = nil
-        }
-
-        func webView(_ webView: WKWebView, didFinish navigation: WKNavigation!) {
-            injectStyles(into: webView)
-            unmuteVideo(in: webView)
-            startPlaybackObserver(for: webView)
-        }
-
-        func webView(
-            _ webView: WKWebView,
-            decidePolicyFor navigationAction: WKNavigationAction
-        ) async -> WKNavigationActionPolicy {
-            guard let url = navigationAction.request.url else { return .allow }
-            let host = url.host?.lowercased() ?? ""
-
-            // Allow YouTube and Google auth navigation
-            if host.contains("youtube.com") || host.contains("youtu.be")
-                || host.contains("google.com") || host.contains("accounts.google.com")
-                || host.contains("consent.youtube.com") {
-                return .allow
-            }
-
-            // Block external navigation
-            return .cancel
-        }
-
-        private func unmuteVideo(in webView: WKWebView) {
-            let script = """
-            (function() {
-                function unmute() {
-                    var video = document.querySelector('video');
-                    if (video) { video.muted = false; }
-                }
-                unmute();
-                var observer = new MutationObserver(function() { unmute(); });
-                observer.observe(document.body, { childList: true, subtree: true });
-                setTimeout(function() { observer.disconnect(); }, 5000);
-            })();
-            """
-            webView.evaluateJavaScript(script, completionHandler: nil)
-        }
-
-        private func injectStyles(into webView: WKWebView) {
-            let css = """
-            * { margin: 0 !important; padding: 0 !important; }
-            body { overflow: hidden !important; background: #000 !important; }
-            #player, .html5-video-player, video {
-                position: fixed !important;
-                top: 0 !important;
-                left: 0 !important;
-                width: 100vw !important;
-                height: 100vh !important;
-                z-index: 999999 !important;
-            }
-            ytd-app, #content, #page-manager, ytd-watch-flexy,
-            #columns, #primary, #primary-inner, #player-container-outer,
-            #player-container-inner, #player-container,
-            #movie_player, .html5-video-container {
-                position: fixed !important;
-                top: 0 !important; left: 0 !important;
-                width: 100vw !important; height: 100vh !important;
-                max-width: none !important; max-height: none !important;
-                min-width: 0 !important; min-height: 0 !important;
-                margin: 0 !important; padding: 0 !important;
-                overflow: hidden !important;
-            }
-            #secondary, #related, #comments, #info, #meta,
-            #above-the-fold, #below, ytd-watch-metadata,
-            #masthead-container, #guide, ytd-masthead,
-            ytd-mini-guide-renderer, #chat, .ytp-chrome-top,
-            .ytp-pause-overlay, .ytp-endscreen-content,
-            .ytp-chrome-bottom, .ytp-gradient-bottom,
-            .ytp-gradient-top, ytd-engagement-panel-section-list-renderer,
-            tp-yt-app-drawer, #description, #actions,
-            ytd-merch-shelf-renderer, ytd-info-panel-content-renderer,
-            .ytd-watch-next-secondary-results-renderer,
-            #chips, #header, .ytd-rich-grid-renderer,
-            ytd-feed-filter-chip-bar-renderer,
-            ytd-compact-promoted-video-renderer,
-            .ytp-ce-element, .ytp-cards-teaser,
-            .ytp-paid-content-overlay, .iv-branding,
-            .ytp-youtube-button, .ytp-watermark,
-            tp-yt-paper-dialog, ytd-popup-container,
-            ytd-consent-bump-v2-lightbox {
-                display: none !important;
-                visibility: hidden !important;
-                height: 0 !important;
-                width: 0 !important;
-                opacity: 0 !important;
-                pointer-events: none !important;
-            }
-            """
-
-            let script = """
-            (function() {
-                var style = document.createElement('style');
-                style.textContent = `\(css)`;
-                document.head.appendChild(style);
-
-                // Re-apply after dynamic content loads
-                var observer = new MutationObserver(function() {
-                    if (!document.getElementById('sakura-yt-style')) {
-                        var s = document.createElement('style');
-                        s.id = 'sakura-yt-style';
-                        s.textContent = `\(css)`;
-                        document.head.appendChild(s);
-                    }
-                });
-                observer.observe(document.body, { childList: true, subtree: true });
-            })();
-            """
-            webView.evaluateJavaScript(script, completionHandler: nil)
-        }
-
-        private func startPlaybackObserver(for webView: WKWebView) {
-            playbackObserver?.invalidate()
-            playbackObserver = Timer.scheduledTimer(withTimeInterval: 0.5, repeats: true) { [weak self] _ in
-                let script = """
-                (function() {
-                    var video = document.querySelector('video');
-                    if (!video) return null;
-                    return { playing: !video.paused, currentTime: video.currentTime, duration: video.duration || 0 };
-                })();
-                """
-                webView.evaluateJavaScript(script) { result, _ in
-                    if let dict = result as? [String: Any] {
-                        DispatchQueue.main.async {
-                            if let playing = dict["playing"] as? Bool {
-                                self?.isPlaying = playing
-                            }
-                            if let time = dict["currentTime"] as? Double {
-                                self?.currentTime = time
-                            }
-                            if let dur = dict["duration"] as? Double, dur > 0 {
-                                self?.duration = dur
-                            }
-                        }
-                    }
-                }
-            }
-        }
-    }
-}
-
-private final class PremiumCheckDelegate: NSObject, WKNavigationDelegate {
-    private let completion: (Bool) -> Void
-    private var hasCompleted = false
-
-    init(completion: @escaping (Bool) -> Void) {
-        self.completion = completion
-        super.init()
-    }
-
-    func webView(_ webView: WKWebView, didFinish navigation: WKNavigation!) {
-        guard !hasCompleted else { return }
-
-        // Check for YouTube Premium by looking for premium membership
-        // indicators in YouTube's page data
-        let script = """
-        (function() {
-            try {
-                var text = document.documentElement.innerHTML;
-                if (text.indexOf('"isPremium":true') !== -1) {
-                    return true;
-                }
-                if (text.indexOf('"hasPaidContent":true') !== -1) {
-                    return true;
-                }
-                if (document.querySelector('[is-premium-member]')) {
-                    return true;
-                }
-                return false;
-            } catch(e) {
-                return false;
-            }
-        })();
-        """
-
-        DispatchQueue.main.asyncAfter(deadline: .now() + 2.0) { [weak self] in
-            guard let self, !self.hasCompleted else { return }
-            webView.evaluateJavaScript(script) { [weak self] result, _ in
-                guard let self, !self.hasCompleted else { return }
-                self.hasCompleted = true
-                self.completion((result as? Bool) == true)
-            }
-        }
-    }
-
-    func webView(_ webView: WKWebView, didFail navigation: WKNavigation!, withError error: Error) {
-        guard !hasCompleted else { return }
-        hasCompleted = true
-        completion(false)
     }
 }
