+++ Shared/XProfileScraper/XProfileNavigationHandler.swift
@@ -1,0 +1,41 @@
+import Foundation
+import WebKit
+
+/// Minimal WKNavigationDelegate that resolves a continuation on load completion.
+@MainActor
+final class XProfileNavigationHandler: NSObject, WKNavigationDelegate {
+
+    private var continuation: CheckedContinuation<Bool, Never>?
+
+    init(continuation: CheckedContinuation<Bool, Never>) {
+        self.continuation = continuation
+    }
+
+    nonisolated func webView(_ webView: WKWebView, didFinish navigation: WKNavigation!) {
+        Task { @MainActor [weak self] in
+            self?.resume(with: true)
+        }
+    }
+
+    nonisolated func webView(_ webView: WKWebView, didFail navigation: WKNavigation!, withError error: Error) {
+        Task { @MainActor [weak self] in
+            self?.resume(with: false)
+        }
+    }
+
+    nonisolated func webView(
+        _ webView: WKWebView,
+        didFailProvisionalNavigation navigation: WKNavigation!,
+        withError error: Error
+    ) {
+        Task { @MainActor [weak self] in
+            self?.resume(with: false)
+        }
+    }
+
+    private func resume(with value: Bool) {
+        guard let continuation else { return }
+        self.continuation = nil
+        continuation.resume(returning: value)
+    }
+}
